refactor(#214): split repositories monolith into package and add connection pooling
Split the 1,229-line db/repositories.py monolith into a proper Python
package under db/repositories/ with one class per file. An __init__.py
re-exports all public names so existing imports continue unchanged.

Also:
- Added AnalyticsRepository (new) for the 5 admin analytics queries; admin
  routes now delegate to it instead of containing raw SQL
- Replaced raw psycopg helpers in chat.py with CallRepository /
  LearningRepository calls; fixed missing logger import
- Added psycopg-pool to api/requirements.txt and wired a ConnectionPool in
  the FastAPI lifespan (dependencies.py falls back to psycopg.connect when
  the pool is absent, keeping tests unaffected)
- Updated test patch target from db.repositories.register_vector to
  db.repositories.embeddings.register_vector

All 115 tests pass.
