+++ api/requirements.txt
@@ -1,6 +1,7 @@
 fastapi>=0.115.0
 uvicorn[standard]>=0.30.0
 psycopg[binary]>=3.2.0
+psycopg-pool>=3.2.0
 python-dotenv>=1.0.0
 PyJWT[crypto]>=2.8.0
 python-multipart>=0.0.9
