+++ db/repositories/analysis.py
@@ -1,169 +1,14 @@
-import uuid
+"""Analysis data repository: read/write NLP results, terms, and synthesis."""
+
 import logging
-from datetime import datetime, timezone, timedelta
+import uuid
 
 import psycopg
-import psycopg.errors
 from pgvector.psycopg import register_vector
-from core.models import CallAnalysis, Competitor, TranscriptChunk
+
+from core.models import CallAnalysis, TranscriptChunk
 
 logger = logging.getLogger(__name__)
-
-class OutdatedSchemaError(Exception):
-    """Exception raised when the database schema is out of date."""
-    pass
-
-
-class RepositoryError(Exception):
-    """Raised when a repository operation fails due to a database error."""
-    pass
-
-REQUIRED_SCHEMA_VERSION = 9
-
-
-def reset_all_data(conn_str: str) -> None:
-    """Delete all application data from the database, preserving the schema."""
-    # Deleting from calls cascades to all dependent tables via ON DELETE CASCADE.
-    # learning_sessions and concept_exercises are not linked to calls so are truncated separately.
-    with psycopg.connect(conn_str) as conn:
-        with conn.cursor() as cur:
-            cur.execute("DELETE FROM learning_sessions")
-            cur.execute("DELETE FROM calls")
-        conn.commit()
-
-
-class SchemaRepository:
-    def __init__(self, conn_str: str):
-        self.conn_str = conn_str
-
-    def get_current_version(self) -> int:
-        """Get the current schema version from the database. Returns 0 if table missing or empty."""
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute("SELECT version FROM schema_version ORDER BY version DESC LIMIT 1")
-                    row = cur.fetchone()
-                    return row[0] if row else 0
-        except psycopg.errors.UndefinedTable:
-            return 0
-        except Exception as e:
-            logger.warning(f"Error checking schema version: {e}")
-            return 0
-
-    def check_health(self) -> tuple[bool, str]:
-        """Check if the database schema is up to date."""
-        current_version = self.get_current_version()
-        if current_version < REQUIRED_SCHEMA_VERSION:
-            if current_version == 0:
-                msg = "Database schema version table is missing. Re-initialize the database."
-            else:
-                msg = f"Database schema is outdated (current: {current_version}, required: {REQUIRED_SCHEMA_VERSION})."
-            return False, f"{msg} Run 'python migrate.py' or './reset_db.sh' to update."
-        return True, "Database schema is up to date."
-
-
-class CallRepository:
-    def __init__(self, conn_str: str):
-        self.conn_str = conn_str
-
-    def get_company_info(self, ticker: str) -> tuple[str, str]:
-        """Return (company_name, industry) for a ticker, or empty strings if not found."""
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        "SELECT company_name, industry FROM calls WHERE ticker = %s LIMIT 1",
-                        (ticker,),
-                    )
-                    row = cur.fetchone()
-                    if row:
-                        return (row[0] or "", row[1] or "")
-        except Exception as e:
-            logger.warning(f"Could not fetch company info for {ticker}: {e}")
-        return ("", "")
-
-    def get_call_date(self, ticker: str):
-        """Return the call_date for a ticker, or None if not set."""
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        "SELECT call_date FROM calls WHERE ticker = %s LIMIT 1",
-                        (ticker,),
-                    )
-                    row = cur.fetchone()
-                    return row[0] if row else None
-        except Exception as e:
-            logger.warning(f"Could not fetch call_date for {ticker}: {e}")
-            return None
-
-    def get_all_calls(self) -> list[tuple[str, str, str | None, str | None]]:
-        """Return (ticker, fiscal_quarter, company_name, call_date) for all stored calls."""
-        calls = []
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        """
-                        SELECT ticker, fiscal_quarter, company_name, call_date
-                        FROM calls
-                        ORDER BY created_at DESC
-                        """
-                    )
-                    calls = cur.fetchall()
-        except Exception as e:
-            logger.warning(f"Could not fetch calls: {e}")
-        return calls
-
-
-class EmbeddingRepository:
-    def __init__(self, conn_str: str):
-        self.conn_str = conn_str
-
-    def fetch_existing_embeddings(self, ticker: str, fiscal_quarter: str) -> dict[str, list[float]]:
-        cache: dict[str, list[float]] = {}
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                register_vector(conn)
-                with conn.cursor() as cur:
-                    cur.execute(
-                        """
-                        SELECT s.text, s.embedding
-                        FROM spans s
-                        JOIN calls c ON s.call_id = c.id
-                        WHERE c.ticker = %s AND c.fiscal_quarter = %s
-                            AND s.embedding IS NOT NULL
-                        """,
-                        (ticker, fiscal_quarter),
-                    )
-                    for text, embedding in cur.fetchall():
-                        if embedding is not None:
-                            cache[text] = embedding.tolist()
-        except Exception as e:
-            logger.warning(f"Could not fetch embedding cache: {e}")
-        return cache
-
-    def search_spans(self, ticker: str, query_vector: list[float], top_k: int = 5) -> list[str]:
-        results = []
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                register_vector(conn)
-                with conn.cursor() as cur:
-                    cur.execute(
-                        """
-                        SELECT s.text
-                        FROM spans s
-                        JOIN calls c ON s.call_id = c.id
-                        WHERE c.ticker = %s
-                        ORDER BY s.embedding <=> %s::vector
-                        LIMIT %s
-                        """,
-                        (ticker, query_vector, top_k),
-                    )
-                    results = [row[0] for row in cur.fetchall()]
-        except Exception as e:
-            logger.warning(f"Could not perform semantic search: {e}")
-        return results
 
 
 class AnalysisRepository:
@@ -602,6 +447,7 @@ class AnalysisRepository:
             return False
 
     def save_analysis(self, result: CallAnalysis) -> None:
+        """Persist a full CallAnalysis to the database."""
         with psycopg.connect(self.conn_str) as conn:
             register_vector(conn)
             with conn.cursor() as cur:
@@ -702,7 +548,7 @@ class AnalysisRepository:
         for t in takeaways:
             if t.id not in all_spans:
                 all_spans[t.id] = t
-                
+
         for span in all_spans.values():
             speaker_id = speaker_ids.get(span.speaker_name)
             cur.execute(
@@ -868,366 +714,3 @@ class AnalysisRepository:
                     gotcha.get("correction") or "",
                 ),
             )
-
-
-_COMPETITOR_CACHE_TTL_DAYS = 30
-
-
-class CompetitorRepository:
-    """Read/write cached competitors for a given ticker."""
-
-    def __init__(self, conn_str: str):
-        self.conn_str = conn_str
-
-    def _get_call_id(self, cur, ticker: str) -> str | None:
-        """Return the call UUID for a ticker, or None if not found."""
-        cur.execute("SELECT id FROM calls WHERE ticker = %s LIMIT 1", (ticker,))
-        row = cur.fetchone()
-        return str(row[0]) if row else None
-
-    def get(self, ticker: str) -> list[Competitor]:
-        """Return cached competitors for a ticker if still within TTL, else empty list."""
-        rows = []
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        """
-                        SELECT co.competitor_name, co.competitor_ticker,
-                               co.description, co.mentioned_in_transcript, co.fetched_at
-                        FROM competitors co
-                        JOIN calls c ON co.call_id = c.id
-                        WHERE c.ticker = %s
-                        ORDER BY co.competitor_name ASC
-                        """,
-                        (ticker,),
-                    )
-                    rows = cur.fetchall()
-        except Exception as e:
-            logger.warning("Could not fetch competitors for %s: %s", ticker, e)
-            return []
-
-        if not rows:
-            return []
-
-        # Check TTL against the oldest fetched_at in the set
-        cutoff = datetime.now(timezone.utc) - timedelta(days=_COMPETITOR_CACHE_TTL_DAYS)
-        fetched_at = rows[0][4]
-        if fetched_at and fetched_at < cutoff:
-            return []
-
-        return [
-            Competitor(
-                name=row[0],
-                ticker=row[1] or "",
-                description=row[2] or "",
-                mentioned_in_transcript=bool(row[3]),
-            )
-            for row in rows
-        ]
-
-    def save(self, ticker: str, competitors: list[Competitor]) -> None:
-        """Delete existing competitors for a ticker and insert the new list."""
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    call_id = self._get_call_id(cur, ticker)
-                    if not call_id:
-                        logger.warning("No call found for ticker %s — skipping competitor save", ticker)
-                        return
-                    cur.execute("DELETE FROM competitors WHERE call_id = %s", (call_id,))
-                    for c in competitors:
-                        cur.execute(
-                            """
-                            INSERT INTO competitors (
-                                call_id, competitor_name, competitor_ticker,
-                                description, mentioned_in_transcript
-                            ) VALUES (%s, %s, %s, %s, %s)
-                            ON CONFLICT (call_id, competitor_name) DO UPDATE SET
-                                competitor_ticker = EXCLUDED.competitor_ticker,
-                                description = EXCLUDED.description,
-                                mentioned_in_transcript = EXCLUDED.mentioned_in_transcript,
-                                fetched_at = now()
-                            """,
-                            (call_id, c.name, c.ticker or None, c.description, c.mentioned_in_transcript),
-                        )
-                conn.commit()
-        except Exception as e:
-            logger.warning("Could not save competitors for %s: %s", ticker, e)
-
-    def delete(self, ticker: str) -> None:
-        """Remove all cached competitors for a ticker (forces re-fetch)."""
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        """
-                        DELETE FROM competitors
-                        USING calls
-                        WHERE competitors.call_id = calls.id AND calls.ticker = %s
-                        """,
-                        (ticker,),
-                    )
-                conn.commit()
-        except Exception as e:
-            logger.warning("Could not delete competitors for %s: %s", ticker, e)
-
-
-SYSTEM_USER_ID = "00000000-0000-0000-0000-000000000001"
-
-
-class LearningRepository:
-    """Read/write Feynman learning sessions and concept exercises."""
-
-    def __init__(self, conn_str: str):
-        self.conn_str = conn_str
-
-    def _get_call_id(self, cur, ticker: str) -> str | None:
-        """Look up the call UUID for a ticker."""
-        cur.execute("SELECT id FROM calls WHERE ticker = %s LIMIT 1", (ticker,))
-        row = cur.fetchone()
-        return str(row[0]) if row else None
-
-    def save_session(
-        self,
-        ticker: str,
-        session_id: str,
-        topic: str,
-        stage: int,
-        messages: list[dict],
-        completed: bool,
-        session_type: str = "feynman",
-        user_id: str = SYSTEM_USER_ID,
-    ) -> bool:
-        """Upsert a learning session. Stores full message history in notes as JSON. Returns True on success."""
-        import json
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    call_id = self._get_call_id(cur, ticker)
-                    if not call_id:
-                        logger.warning(f"No call found for ticker {ticker}, skipping session save")
-                        return False
-                    notes = json.dumps({"topic": topic, "stage": stage, "messages": messages, "type": session_type})
-                    completed_at_expr = "now()" if completed else "NULL"
-                    cur.execute(
-                        f"""
-                        INSERT INTO learning_sessions (id, user_id, call_id, notes, completed_at)
-                        VALUES (%s, %s::uuid, %s, %s, {completed_at_expr})
-                        ON CONFLICT (id) DO UPDATE SET
-                            notes = EXCLUDED.notes,
-                            completed_at = COALESCE(learning_sessions.completed_at, EXCLUDED.completed_at)
-                        """,
-                        (session_id, user_id, call_id, notes),
-                    )
-                    if completed:
-                        teaching_note = next(
-                            (m["content"] for m in reversed(messages)
-                             if m.get("role") == "assistant" and m.get("feynman_stage") == 5),
-                            None,
-                        )
-                        cur.execute(
-                            """
-                            INSERT INTO concept_exercises (session_id, concept_label, ai_critique)
-                            SELECT %s, %s, %s
-                            WHERE NOT EXISTS (SELECT 1 FROM concept_exercises WHERE session_id = %s)
-                            """,
-                            (session_id, topic, teaching_note, session_id),
-                        )
-                conn.commit()
-            return True
-        except Exception as e:
-            logger.warning(f"Could not save learning session: {e}")
-            return False
-
-    def get_sessions_for_ticker(self, ticker: str, user_id: str | None = None) -> list[dict]:
-        """Return all sessions for a ticker, newest first. Each dict has: id, topic, stage, completed, teaching_note, started_at.
-
-        When user_id is provided only sessions belonging to that user are returned.
-        """
-        import json
-        rows = []
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    if user_id is not None:
-                        cur.execute(
-                            """
-                            SELECT ls.id, ls.notes, ls.completed_at, ls.started_at,
-                                   ce.ai_critique
-                            FROM learning_sessions ls
-                            JOIN calls c ON ls.call_id = c.id
-                            LEFT JOIN concept_exercises ce ON ce.session_id = ls.id
-                            WHERE c.ticker = %s AND ls.user_id = %s::uuid
-                            ORDER BY ls.started_at DESC
-                            """,
-                            (ticker, user_id),
-                        )
-                    else:
-                        cur.execute(
-                            """
-                            SELECT ls.id, ls.notes, ls.completed_at, ls.started_at,
-                                   ce.ai_critique
-                            FROM learning_sessions ls
-                            JOIN calls c ON ls.call_id = c.id
-                            LEFT JOIN concept_exercises ce ON ce.session_id = ls.id
-                            WHERE c.ticker = %s
-                            ORDER BY ls.started_at DESC
-                            """,
-                            (ticker,),
-                        )
-                    for row in cur.fetchall():
-                        session_id, notes_json, completed_at, started_at, teaching_note = row
-                        notes = json.loads(notes_json) if notes_json else {}
-                        rows.append({
-                            "id": str(session_id),
-                            "topic": notes.get("topic", "Unknown topic"),
-                            "stage": notes.get("stage", 1),
-                            "messages": notes.get("messages", []),
-                            "completed": completed_at is not None,
-                            "teaching_note": teaching_note,
-                            "started_at": started_at,
-                            "session_type": notes.get("type", "feynman"),
-                        })
-        except Exception as e:
-            logger.warning(f"Could not fetch sessions for ticker {ticker}: {e}")
-        return rows
-
-    def get_session_by_id(self, session_id: str, user_id: str) -> dict | None:
-        """Return session data for the given session_id if it belongs to user_id.
-
-        Returns None if not found. Raises ValueError if the session belongs to a different user.
-        """
-        import json
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        "SELECT notes, user_id FROM learning_sessions WHERE id = %s LIMIT 1",
-                        (session_id,),
-                    )
-                    row = cur.fetchone()
-            if not row:
-                return None
-            notes_json, owner_id = row
-            if str(owner_id) != user_id:
-                raise ValueError(f"Session {session_id!r} belongs to a different user")
-            return json.loads(notes_json) if notes_json else {}
-        except ValueError:
-            raise
-        except Exception as e:
-            logger.warning(f"Could not fetch session {session_id}: {e}")
-            return None
-
-    def get_learning_stats(self) -> dict:
-        """Return overall learning stats: tickers_studied, total_sessions, completed_sessions."""
-        stats = {"tickers_studied": 0, "total_sessions": 0, "completed_sessions": 0}
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        """
-                        SELECT COUNT(DISTINCT c.ticker),
-                               COUNT(ls.id),
-                               SUM(CASE WHEN ls.completed_at IS NOT NULL THEN 1 ELSE 0 END)
-                        FROM learning_sessions ls
-                        JOIN calls c ON ls.call_id = c.id
-                        """
-                    )
-                    row = cur.fetchone()
-                    if row:
-                        stats["tickers_studied"] = row[0] or 0
-                        stats["total_sessions"] = row[1] or 0
-                        stats["completed_sessions"] = int(row[2] or 0)
-        except Exception as e:
-            logger.warning(f"Could not fetch learning stats: {e}")
-        return stats
-
-    def get_ticker_session_counts(self) -> list[tuple[str, int, int]]:
-        """Return (ticker, total_sessions, completed_sessions) for all tickers that have learning history."""
-        rows = []
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        """
-                        SELECT c.ticker,
-                               COUNT(ls.id),
-                               SUM(CASE WHEN ls.completed_at IS NOT NULL THEN 1 ELSE 0 END)
-                        FROM learning_sessions ls
-                        JOIN calls c ON ls.call_id = c.id
-                        GROUP BY c.ticker
-                        ORDER BY COUNT(ls.id) DESC
-                        """
-                    )
-                    rows = [(r[0], r[1], int(r[2] or 0)) for r in cur.fetchall()]
-        except Exception as e:
-            logger.warning(f"Could not fetch ticker session counts: {e}")
-        return rows
-
-
-class ProgressRepository:
-    """Read/write per-step completion tracking for the 6-step learning path."""
-
-    def __init__(self, conn_str: str):
-        self.conn_str = conn_str
-
-    def mark_step_viewed(self, ticker: str, step_number: int) -> None:
-        """Record that a step has been read for a given ticker. Idempotent."""
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute("SELECT id FROM calls WHERE ticker = %s LIMIT 1", (ticker,))
-                    row = cur.fetchone()
-                    if not row:
-                        return
-                    call_id = str(row[0])
-                    cur.execute(
-                        """
-                        INSERT INTO transcript_progress (call_id, step_number)
-                        VALUES (%s, %s)
-                        ON CONFLICT (call_id, step_number) DO NOTHING
-                        """,
-                        (call_id, step_number),
-                    )
-                conn.commit()
-        except Exception as e:
-            logger.warning(f"Could not mark step {step_number} as viewed for {ticker}: {e}")
-
-    def get_completed_steps(self, ticker: str) -> set[int]:
-        """Return the set of step numbers the user has marked as read for a transcript."""
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        """
-                        SELECT tp.step_number
-                        FROM transcript_progress tp
-                        JOIN calls c ON tp.call_id = c.id
-                        WHERE c.ticker = %s
-                        """,
-                        (ticker,),
-                    )
-                    return {row[0] for row in cur.fetchall()}
-        except Exception as e:
-            logger.warning(f"Could not fetch progress for {ticker}: {e}")
-        return set()
-
-    def get_all_step_counts(self) -> list[tuple[str, int]]:
-        """Return (ticker, steps_completed) for all transcripts with any step progress."""
-        try:
-            with psycopg.connect(self.conn_str) as conn:
-                with conn.cursor() as cur:
-                    cur.execute(
-                        """
-                        SELECT c.ticker, COUNT(tp.step_number)
-                        FROM transcript_progress tp
-                        JOIN calls c ON tp.call_id = c.id
-                        GROUP BY c.ticker
-                        ORDER BY COUNT(tp.step_number) DESC
-                        """
-                    )
-                    return [(row[0], row[1]) for row in cur.fetchall()]
-        except Exception as e:
-            logger.warning(f"Could not fetch step counts: {e}")
-        return []
